Fixed heart and bookmark icon not staying clicked when going to home page.

+++ 4370_p2_micro_blog-main/src/main/java/uga/menik/cs4370/services/MakePostService.java
@@ -365,13 +365,14 @@ public class MakePostService {
         } else {
             // unliking the post
 
-            final String sqlString = "delete from heart where postId = ?";
+            final String sqlString = "delete from heart where postId = ? and userId = ?";
 
             String userId = userService.getLoggedInUser().getUserId();
     
             try (Connection conn = dataSource.getConnection(); PreparedStatement pstmt = conn.prepareStatement(sqlString)) {
     
                 pstmt.setString(1, postId);
+                pstmt.setString(2, userId);
     
                 int rowsChanged = pstmt.executeUpdate();
     
@@ -421,13 +422,14 @@ public class MakePostService {
         } else {
             // unliking the post
 
-            final String sqlString = "delete from bookmark where postId = ?";
+            final String sqlString = "delete from bookmark where postId = ? and userId = ?";
 
             String userId = userService.getLoggedInUser().getUserId();
     
             try (Connection conn = dataSource.getConnection(); PreparedStatement pstmt = conn.prepareStatement(sqlString)) {
     
                 pstmt.setString(1, postId);
+                pstmt.setString(2, userId);
     
                 int rowsChanged = pstmt.executeUpdate();
